refactor(netty): serialize attachment size as 4-byte int
Use writeInt/readInt for the attachment size field instead of
writeByte/readByte. This keeps the size field consistent with the
other length fields in the header (crcCode, length, keySize, body
length are all 4-byte ints) and removes the need for the & 0xff
unsigned conversion on the decoder side.

The frame length grows by 3 bytes per message (e.g. heartbeat
frames from 23 to 26 bytes, login response from 98 to 101 bytes),
which is negligible relative to the total message size.

+++ src/main/java/net/thewesthill/protocol/netty/codec/NettyMessageEncoder.java
@@ -31,7 +31,7 @@ public class NettyMessageEncoder extends MessageToMessageEncoder<NettyMessage> {
     sendBuf.writeLong(msg.getHeader().getSessionId());
     sendBuf.writeByte(msg.getHeader().getType());
     sendBuf.writeByte(msg.getHeader().getPriority());
-    sendBuf.writeByte(msg.getHeader().getAttachment().size());
+    sendBuf.writeInt(msg.getHeader().getAttachment().size());
     String key = null;
     byte[] keyArray = null;
     Object value = null;
